+++ .js/plugins/english/nobleMTL.js
@@ -39,32 +39,32 @@ Object.defineProperty(exports, "__esModule", { value: true });
 var cheerio_1 = require("cheerio");
 var fetch_1 = require("@libs/fetch");
 var filterInputs_1 = require("@libs/filterInputs");
-var KolNovel = /** @class */ (function () {
-    function KolNovel() {
-        this.id = "kolnovel";
-        this.name = "KolNovel";
-        this.icon = "multisrc/wpmangastream/icons/kolnovel.png";
-        this.site = "https://kolnovel.lol/";
+var NobleMTL = /** @class */ (function () {
+    function NobleMTL() {
+        this.id = "noblemtl";
+        this.name = "NobleMTL";
+        this.icon = "src/en/noblemtl/icon.png";
+        this.site = "https://noblemtl.com/";
         this.version = "1.0.0";
         this.userAgent = "";
         this.cookieString = "";
         this.filters = [
             {
                 key: "order",
-                label: "ترتيب حسب",
+                label: "Sort By",
                 values: [
-                    { label: "الإعداد الأولي", value: "" },
+                    { label: "Default", value: "" },
                     { label: "A-Z", value: "title" },
                     { label: "Z-A", value: "titlereverse" },
-                    { label: "أخر التحديثات", value: "update" },
-                    { label: "أخر ما تم إضافته", value: "latest" },
-                    { label: "الرائجة", value: "popular" },
+                    { label: "Latest Update", value: "update" },
+                    { label: "Latest Added", value: "latest" },
+                    { label: "Popular", value: "popular" },
                 ],
                 inputType: filterInputs_1.FilterInputs.Picker,
             },
             {
                 key: "status",
-                label: "الحالة",
+                label: "Status",
                 values: [
                     { label: "All", value: "" },
                     { label: "Ongoing", value: "ongoing" },
@@ -75,65 +75,113 @@ var KolNovel = /** @class */ (function () {
             },
             {
                 key: "type",
-                label: "النوع",
+                label: "Type",
                 values: [
-                    { label: "إنجليزية", value: "english" },
-                    { label: "روايةلايت", value: "light-novel" },
-                    { label: "روايةويب", value: "web-novel" },
-                    { label: "صينية", value: "chinese" },
-                    { label: "عربية", value: "arabic" },
-                    { label: "كورية", value: "korean" },
-                    { label: "يابانية", value: "japanese" },
+                    { label: "Chinese novel", value: "chinese-novel" },
+                    { label: "habyeol", value: "habyeol" },
+                    { label: "korean novel", value: "korean-novel" },
+                    { label: "Web Novel", value: "web-novel" },
+                    { label: "삼심", value: "%ec%82%bc%ec%8b%ac" },
+                    { label: "호곡", value: "%ed%98%b8%ea%b3%a1" },
                 ],
                 inputType: filterInputs_1.FilterInputs.Checkbox,
             },
             {
                 key: "genres",
-                label: "تصنيف",
+                label: "Genres",
                 values: [
+                    { label: "A.I", value: "a.i" },
+                    { label: "Academy", value: "academy" },
+                    { label: "Action", value: "action" },
+                    { label: "Adult", value: "adult" },
+                    { label: "Adventure", value: "adventure" },
+                    { label: "Alternative History", value: "alternative-history" },
+                    { label: "Another World", value: "another-world" },
+                    { label: "Apocalypse", value: "apocalypse" },
+                    { label: "Bromance", value: "bromance" },
+                    { label: "Comedy", value: "comedy" },
+                    { label: "Dark fantasy", value: "dark-fantasy" },
+                    { label: "Demons", value: "demons" },
+                    { label: "Drama", value: "drama" },
+                    { label: "Dystopia", value: "dystopia" },
+                    { label: "Ecchi", value: "ecchi" },
+                    { label: "Entertainment", value: "entertainment" },
+                    { label: "Exhaustion", value: "exhaustion" },
+                    { label: "Fanfiction", value: "fanfiction" },
+                    { label: "fantasy", value: "fantasy" },
+                    { label: "finance", value: "finance" },
+                    { label: "Full color", value: "full-color" },
+                    { label: "Game", value: "game" },
+                    { label: "Gender Bender", value: "gender-bender" },
+                    { label: "Genius", value: "genius" },
+                    { label: "Harem", value: "harem" },
+                    { label: "Hero", value: "hero" },
+                    { label: "Historical", value: "historical" },
+                    { label: "Hunter", value: "hunter" },
+                    { label: "korean novel", value: "korean-novel" },
+                    { label: "Light Novel", value: "light-novel" },
+                    {
+                        label: "List Adventure Manga Genres",
+                        value: "list-adventure-manga-genres",
+                    },
+                    { label: "Long Strip", value: "long-strip" },
+                    { label: "Love comedy", value: "love-comedy" },
+                    { label: "magic", value: "magic" },
+                    { label: "Manhua", value: "manhua" },
+                    { label: "Martial Arts", value: "martial-arts" },
+                    { label: "Mature", value: "mature" },
+                    { label: "Medieval", value: "medieval" },
+                    { label: "Misunderstanding", value: "misunderstanding" },
+                    { label: "Modern", value: "modern" },
+                    { label: "modern fantasy", value: "modern-fantasy" },
+                    { label: "music", value: "music" },
+                    { label: "Mystery", value: "mystery" },
+                    { label: "Necromancy", value: "necromancy" },
+                    { label: "No Romance", value: "no-romance" },
+                    { label: "NTL", value: "ntl" },
+                    { label: "o", value: "o" },
+                    { label: "Obsession", value: "obsession" },
+                    { label: "Politics", value: "politics" },
+                    { label: "Possession", value: "possession" },
+                    { label: "Programming", value: "programming" },
+                    { label: "Psychological", value: "psychological" },
+                    { label: "Pure Love", value: "pure-love" },
+                    { label: "Redemption", value: "redemption" },
+                    { label: "Regression", value: "regression" },
+                    { label: "Regret", value: "regret" },
+                    { label: "Reincarnation", value: "reincarnation" },
+                    { label: "Revenge", value: "revenge" },
+                    { label: "Romance", value: "romance" },
+                    { label: "Romance Fanrasy", value: "romance-fanrasy" },
+                    { label: "Salvation", value: "salvation" },
+                    { label: "School Life", value: "school-life" },
+                    { label: "Sci-fi", value: "sci-fi" },
+                    { label: "Science fiction", value: "science-fiction" },
+                    { label: "Seinen", value: "seinen" },
+                    { label: "Shounen", value: "shounen" },
+                    { label: "Slice of Life", value: "slice-of-life" },
+                    { label: "Soft yandere", value: "soft-yandere" },
+                    { label: "Sports", value: "sports" },
+                    { label: "Supernatural", value: "supernatural" },
+                    { label: "Survival", value: "survival" },
+                    { label: "system", value: "system" },
+                    { label: "Time limit", value: "time-limit" },
+                    { label: "Tragedy", value: "tragedy" },
+                    { label: "Transmigration", value: "transmigration" },
+                    { label: "TS", value: "ts" },
+                    { label: "Tsundere", value: "tsundere" },
+                    { label: "Unique", value: "unique" },
+                    { label: "Wholesome", value: "wholesome" },
                     { label: "Wuxia", value: "wuxia" },
-                    { label: "Xianxia", value: "xianxia" },
-                    { label: "XUANHUAN", value: "xuanhuan" },
-                    { label: "أكشن", value: "action" },
-                    { label: "إثارة", value: "excitement" },
-                    { label: "إنتقالالىعالمأخر", value: "isekai" },
-                    { label: "إيتشي", value: "etchi" },
-                    { label: "الخيالالعلمي", value: "sci-fi" },
-                    { label: "بوليسي", value: "policy" },
-                    { label: "تاريخي", value: "historical" },
-                    { label: "تحقيقات", value: "%d8%aa%d8%ad%d9%82%d9%8a%d9%82" },
-                    { label: "تقمصشخصيات", value: "rpg" },
-                    { label: "جريمة", value: "crime" },
-                    { label: "جوسى", value: "josei" },
-                    { label: "حريم", value: "harem" },
-                    { label: "حياةمدرسية", value: "school-life" },
-                    { label: "خيالي(فانتازيا)", value: "fantasy" },
-                    { label: "دراما", value: "drama" },
-                    { label: "رعب", value: "horror" },
-                    { label: "رومانسي", value: "romantic" },
-                    { label: "سحر", value: "magic" },
-                    { label: "سينن", value: "senen" },
-                    { label: "شريحةمنالحياة", value: "slice-of-life" },
-                    { label: "شوجو", value: "shojo" },
-                    { label: "شونين", value: "shonen" },
-                    { label: "طبي", value: "medical" },
-                    { label: "ظواهرخارقةللطبيعة", value: "supernatural" },
-                    { label: "غموض", value: "mysteries" },
-                    { label: "فنونالقتال", value: "martial-arts" },
-                    { label: "قوىخارقة", value: "superpower" },
-                    { label: "كوميدي", value: "comedy" },
-                    { label: "مأساوي", value: "tragedy" },
-                    { label: "مابعدالكارثة", value: "after-the-disaster" },
-                    { label: "مغامرة", value: "adventure" },
-                    { label: "ميكا", value: "mechanical" },
-                    { label: "ناضج", value: "mature" },
-                    { label: "نفسي", value: "psychological" },
+                    { label: "Xuanhuan", value: "xuanhuan" },
+                    { label: "Yandere", value: "yandere" },
+                    { label: "Yuri", value: "yuri" },
                 ],
                 inputType: filterInputs_1.FilterInputs.Checkbox,
             },
         ];
     }
-    KolNovel.prototype.popularNovels = function (pageNo, _a) {
+    NobleMTL.prototype.popularNovels = function (pageNo, _a) {
         var filters = _a.filters;
         return __awaiter(this, void 0, void 0, function () {
             var link, headers, body, loadedCheerio, novels;
@@ -161,11 +209,11 @@ var KolNovel = /** @class */ (function () {
                         body = _b.sent();
                         loadedCheerio = (0, cheerio_1.load)(body);
                         novels = [];
-                        loadedCheerio("article.maindet").each(function () {
-                            var novelName = loadedCheerio(this).find("h2").text();
+                        loadedCheerio("article.bs").each(function () {
+                            var novelName = loadedCheerio(this).find(".ntitle").text().trim();
                             var image = loadedCheerio(this).find("img");
                             var novelCover = image.attr("data-src") || image.attr("src");
-                            var novelUrl = loadedCheerio(this).find("h2 a").attr("href");
+                            var novelUrl = loadedCheerio(this).find("a").attr("href");
                             if (!novelUrl)
                                 return;
                             var novel = {
@@ -180,7 +228,8 @@ var KolNovel = /** @class */ (function () {
             });
         });
     };
-    KolNovel.prototype.parseNovelAndChapters = function (novelUrl) {
+    ;
+    NobleMTL.prototype.parseNovelAndChapters = function (novelUrl) {
         return __awaiter(this, void 0, void 0, function () {
             var url, headers, result, body, loadedCheerio, novel, chapter;
             return __generator(this, function (_a) {
@@ -191,7 +240,7 @@ var KolNovel = /** @class */ (function () {
                         if (this.cookieString) {
                             headers.append("cookie", this.cookieString);
                         }
-                        return [4 /*yield*/, fetch(url, { headers: headers })];
+                        return [4 /*yield*/, (0, fetch_1.fetchApi)(url, { headers: headers })];
                     case 1:
                         result = _a.sent();
                         return [4 /*yield*/, result.text()];
@@ -202,27 +251,28 @@ var KolNovel = /** @class */ (function () {
                             url: url,
                             chapters: [],
                         };
-                        novel.name = loadedCheerio("h1.entry-title").text();
+                        novel.name = loadedCheerio(".entry-title").text();
                         novel.cover =
                             loadedCheerio("img.wp-post-image").attr("data-src") ||
                                 loadedCheerio("img.wp-post-image").attr("src");
-                        loadedCheerio("div.serl:nth-child(3) > span").each(function () {
-                            var detailName = loadedCheerio(this).text().trim();
-                            var detail = loadedCheerio(this).next().text().trim();
+                        loadedCheerio("div.spe > span").each(function () {
+                            var detailName = loadedCheerio(this).find("b").text().trim();
+                            var detail = loadedCheerio(this).find('b').remove().end().text().trim();
                             switch (detailName) {
-                                case "الكاتب":
-                                case "Author":
+                                case "Author:":
                                     novel.author = detail;
+                                    break;
+                                case "Status:":
+                                    novel.status = detail;
                                     break;
                             }
                         });
-                        novel.status = loadedCheerio("div.sertostat > span").attr("class");
-                        novel.genres = loadedCheerio(".sertogenre")
-                            .children("a")
-                            .map(function (i, el) { return loadedCheerio(el).text(); })
-                            .toArray()
-                            .join(",");
-                        novel.summary = loadedCheerio(".sersys")
+                        novel.genres = loadedCheerio(".genxed")
+                            .text()
+                            .trim()
+                            .replace(/\s/g, ",");
+                        loadedCheerio('div[itemprop="description"]  h3,p.a,strong').remove();
+                        novel.summary = loadedCheerio('div[itemprop="description"]')
                             .find("br")
                             .replaceWith("\n")
                             .end()
@@ -253,35 +303,30 @@ var KolNovel = /** @class */ (function () {
             });
         });
     };
-    KolNovel.prototype.parseChapter = function (chapterUrl) {
-        var _a;
+    NobleMTL.prototype.parseChapter = function (chapterUrl) {
         return __awaiter(this, void 0, void 0, function () {
-            var headers, result, body, loadedCheerio, ignore, chapterText;
-            return __generator(this, function (_b) {
-                switch (_b.label) {
+            var headers, result, body, loadedCheerio, chapterText;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
                     case 0:
                         headers = new Headers();
                         if (this.cookieString) {
                             headers.append("cookie", this.cookieString);
                         }
-                        return [4 /*yield*/, fetch(chapterUrl, { headers: headers })];
+                        return [4 /*yield*/, (0, fetch_1.fetchApi)(chapterUrl, { headers: headers })];
                     case 1:
-                        result = _b.sent();
+                        result = _a.sent();
                         return [4 /*yield*/, result.text()];
                     case 2:
-                        body = _b.sent();
+                        body = _a.sent();
                         loadedCheerio = (0, cheerio_1.load)(body);
-                        loadedCheerio('.epcontent > div, i').remove();
-                        ignore = loadedCheerio('article > style').text().trim().split(',');
-                        ignore.push.apply(ignore, (((_a = ignore.pop()) === null || _a === void 0 ? void 0 : _a.match(/^\.\w+/)) || []));
-                        ignore.map(function (tag) { return loadedCheerio("p".concat(tag)).remove(); });
-                        chapterText = loadedCheerio('.epcontent').html() || "";
+                        chapterText = loadedCheerio(".epcontent").html() || "";
                         return [2 /*return*/, chapterText];
                 }
             });
         });
     };
-    KolNovel.prototype.searchNovels = function (searchTerm, pageNo) {
+    NobleMTL.prototype.searchNovels = function (searchTerm, pageNo) {
         return __awaiter(this, void 0, void 0, function () {
             var url, headers, result, body, loadedCheerio, novels;
             return __generator(this, function (_a) {
@@ -300,11 +345,10 @@ var KolNovel = /** @class */ (function () {
                         body = _a.sent();
                         loadedCheerio = (0, cheerio_1.load)(body);
                         novels = [];
-                        loadedCheerio("article.maindet").each(function () {
-                            var novelName = loadedCheerio(this).find("h2").text();
-                            var image = loadedCheerio(this).find("img");
-                            var novelCover = image.attr("data-src") || image.attr("src");
-                            var novelUrl = loadedCheerio(this).find("h2 a").attr("href");
+                        loadedCheerio("article.bs").each(function () {
+                            var novelName = loadedCheerio(this).find(".ntitle").text().trim();
+                            var novelCover = loadedCheerio(this).find("img").attr("src");
+                            var novelUrl = loadedCheerio(this).find("a").attr("href");
                             if (!novelUrl)
                                 return;
                             novels.push({
@@ -318,7 +362,7 @@ var KolNovel = /** @class */ (function () {
             });
         });
     };
-    KolNovel.prototype.fetchImage = function (url) {
+    NobleMTL.prototype.fetchImage = function (url) {
         return __awaiter(this, void 0, void 0, function () {
             return __generator(this, function (_a) {
                 switch (_a.label) {
@@ -328,6 +372,6 @@ var KolNovel = /** @class */ (function () {
             });
         });
     };
-    return KolNovel;
+    return NobleMTL;
 }());
-exports.default = new KolNovel();
+exports.default = new NobleMTL();
