+++ .js/plugins/english/pandaMTL.js
@@ -39,32 +39,32 @@ Object.defineProperty(exports, "__esModule", { value: true });
 var cheerio_1 = require("cheerio");
 var fetch_1 = require("@libs/fetch");
 var filterInputs_1 = require("@libs/filterInputs");
-var KolNovel = /** @class */ (function () {
-    function KolNovel() {
-        this.id = "kolnovel";
-        this.name = "KolNovel";
-        this.icon = "multisrc/wpmangastream/icons/kolnovel.png";
-        this.site = "https://kolnovel.lol/";
+var PandaMTL = /** @class */ (function () {
+    function PandaMTL() {
+        this.id = "pandamtl";
+        this.name = "PandaMTL";
+        this.icon = "src/en/wordpress/icon.png";
+        this.site = "https://pandamtl.com/";
         this.version = "1.0.0";
         this.userAgent = "";
         this.cookieString = "";
         this.filters = [
             {
                 key: "order",
-                label: "ترتيب حسب",
+                label: "Sort By",
                 values: [
-                    { label: "الإعداد الأولي", value: "" },
+                    { label: "Default", value: "" },
                     { label: "A-Z", value: "title" },
                     { label: "Z-A", value: "titlereverse" },
-                    { label: "أخر التحديثات", value: "update" },
-                    { label: "أخر ما تم إضافته", value: "latest" },
-                    { label: "الرائجة", value: "popular" },
+                    { label: "Latest Update", value: "update" },
+                    { label: "Latest Added", value: "latest" },
+                    { label: "Popular", value: "popular" },
                 ],
                 inputType: filterInputs_1.FilterInputs.Picker,
             },
             {
                 key: "status",
-                label: "الحالة",
+                label: "Status",
                 values: [
                     { label: "All", value: "" },
                     { label: "Ongoing", value: "ongoing" },
@@ -75,65 +75,42 @@ var KolNovel = /** @class */ (function () {
             },
             {
                 key: "type",
-                label: "النوع",
+                label: "Type",
                 values: [
-                    { label: "إنجليزية", value: "english" },
-                    { label: "روايةلايت", value: "light-novel" },
-                    { label: "روايةويب", value: "web-novel" },
-                    { label: "صينية", value: "chinese" },
-                    { label: "عربية", value: "arabic" },
-                    { label: "كورية", value: "korean" },
-                    { label: "يابانية", value: "japanese" },
+                    { label: "Light Novel (KR)", value: "light-novel-kr" },
+                    { label: "Web Novel", value: "web-novel" },
                 ],
                 inputType: filterInputs_1.FilterInputs.Checkbox,
             },
             {
                 key: "genres",
-                label: "تصنيف",
+                label: "Genres",
                 values: [
-                    { label: "Wuxia", value: "wuxia" },
-                    { label: "Xianxia", value: "xianxia" },
-                    { label: "XUANHUAN", value: "xuanhuan" },
-                    { label: "أكشن", value: "action" },
-                    { label: "إثارة", value: "excitement" },
-                    { label: "إنتقالالىعالمأخر", value: "isekai" },
-                    { label: "إيتشي", value: "etchi" },
-                    { label: "الخيالالعلمي", value: "sci-fi" },
-                    { label: "بوليسي", value: "policy" },
-                    { label: "تاريخي", value: "historical" },
-                    { label: "تحقيقات", value: "%d8%aa%d8%ad%d9%82%d9%8a%d9%82" },
-                    { label: "تقمصشخصيات", value: "rpg" },
-                    { label: "جريمة", value: "crime" },
-                    { label: "جوسى", value: "josei" },
-                    { label: "حريم", value: "harem" },
-                    { label: "حياةمدرسية", value: "school-life" },
-                    { label: "خيالي(فانتازيا)", value: "fantasy" },
-                    { label: "دراما", value: "drama" },
-                    { label: "رعب", value: "horror" },
-                    { label: "رومانسي", value: "romantic" },
-                    { label: "سحر", value: "magic" },
-                    { label: "سينن", value: "senen" },
-                    { label: "شريحةمنالحياة", value: "slice-of-life" },
-                    { label: "شوجو", value: "shojo" },
-                    { label: "شونين", value: "shonen" },
-                    { label: "طبي", value: "medical" },
-                    { label: "ظواهرخارقةللطبيعة", value: "supernatural" },
-                    { label: "غموض", value: "mysteries" },
-                    { label: "فنونالقتال", value: "martial-arts" },
-                    { label: "قوىخارقة", value: "superpower" },
-                    { label: "كوميدي", value: "comedy" },
-                    { label: "مأساوي", value: "tragedy" },
-                    { label: "مابعدالكارثة", value: "after-the-disaster" },
-                    { label: "مغامرة", value: "adventure" },
-                    { label: "ميكا", value: "mechanical" },
-                    { label: "ناضج", value: "mature" },
-                    { label: "نفسي", value: "psychological" },
+                    { label: "Action", value: "action" },
+                    { label: "Adult", value: "adult" },
+                    { label: "Adventure", value: "adventure" },
+                    { label: "Comedy", value: "comedy" },
+                    { label: "Ecchi", value: "ecchi" },
+                    { label: "Fantasy", value: "fantasy" },
+                    { label: "Harem", value: "harem" },
+                    { label: "Josei", value: "josei" },
+                    { label: "Martial Arts", value: "martial-arts" },
+                    { label: "Mature", value: "mature" },
+                    { label: "Romance", value: "romance" },
+                    { label: "School Life", value: "school-life" },
+                    { label: "Sci-fi", value: "sci-fi" },
+                    { label: "Seinen", value: "seinen" },
+                    { label: "Slice of Life", value: "slice-of-life" },
+                    { label: "Smut", value: "smut" },
+                    { label: "Sports", value: "sports" },
+                    { label: "Supernatural", value: "supernatural" },
+                    { label: "Tragedy", value: "tragedy" },
                 ],
                 inputType: filterInputs_1.FilterInputs.Checkbox,
             },
         ];
     }
-    KolNovel.prototype.popularNovels = function (pageNo, _a) {
+    PandaMTL.prototype.popularNovels = function (pageNo, _a) {
         var filters = _a.filters;
         return __awaiter(this, void 0, void 0, function () {
             var link, headers, body, loadedCheerio, novels;
@@ -180,7 +157,7 @@ var KolNovel = /** @class */ (function () {
             });
         });
     };
-    KolNovel.prototype.parseNovelAndChapters = function (novelUrl) {
+    PandaMTL.prototype.parseNovelAndChapters = function (novelUrl) {
         return __awaiter(this, void 0, void 0, function () {
             var url, headers, result, body, loadedCheerio, novel, chapter;
             return __generator(this, function (_a) {
@@ -191,7 +168,7 @@ var KolNovel = /** @class */ (function () {
                         if (this.cookieString) {
                             headers.append("cookie", this.cookieString);
                         }
-                        return [4 /*yield*/, fetch(url, { headers: headers })];
+                        return [4 /*yield*/, (0, fetch_1.fetchApi)(url, { headers: headers })];
                     case 1:
                         result = _a.sent();
                         return [4 /*yield*/, result.text()];
@@ -206,17 +183,16 @@ var KolNovel = /** @class */ (function () {
                         novel.cover =
                             loadedCheerio("img.wp-post-image").attr("data-src") ||
                                 loadedCheerio("img.wp-post-image").attr("src");
-                        loadedCheerio("div.serl:nth-child(3) > span").each(function () {
+                        loadedCheerio(".serl:nth-child(3) > span").each(function () {
                             var detailName = loadedCheerio(this).text().trim();
                             var detail = loadedCheerio(this).next().text().trim();
                             switch (detailName) {
-                                case "الكاتب":
                                 case "Author":
                                     novel.author = detail;
                                     break;
                             }
                         });
-                        novel.status = loadedCheerio("div.sertostat > span").attr("class");
+                        novel.status = loadedCheerio(".sertostat > span").attr("class");
                         novel.genres = loadedCheerio(".sertogenre")
                             .children("a")
                             .map(function (i, el) { return loadedCheerio(el).text(); })
@@ -253,35 +229,30 @@ var KolNovel = /** @class */ (function () {
             });
         });
     };
-    KolNovel.prototype.parseChapter = function (chapterUrl) {
-        var _a;
+    PandaMTL.prototype.parseChapter = function (chapterUrl) {
         return __awaiter(this, void 0, void 0, function () {
-            var headers, result, body, loadedCheerio, ignore, chapterText;
-            return __generator(this, function (_b) {
-                switch (_b.label) {
+            var headers, result, body, loadedCheerio, chapterText;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
                     case 0:
                         headers = new Headers();
                         if (this.cookieString) {
                             headers.append("cookie", this.cookieString);
                         }
-                        return [4 /*yield*/, fetch(chapterUrl, { headers: headers })];
+                        return [4 /*yield*/, (0, fetch_1.fetchApi)(chapterUrl, { headers: headers })];
                     case 1:
-                        result = _b.sent();
+                        result = _a.sent();
                         return [4 /*yield*/, result.text()];
                     case 2:
-                        body = _b.sent();
+                        body = _a.sent();
                         loadedCheerio = (0, cheerio_1.load)(body);
-                        loadedCheerio('.epcontent > div, i').remove();
-                        ignore = loadedCheerio('article > style').text().trim().split(',');
-                        ignore.push.apply(ignore, (((_a = ignore.pop()) === null || _a === void 0 ? void 0 : _a.match(/^\.\w+/)) || []));
-                        ignore.map(function (tag) { return loadedCheerio("p".concat(tag)).remove(); });
-                        chapterText = loadedCheerio('.epcontent').html() || "";
+                        chapterText = loadedCheerio(".epcontent").html() || "";
                         return [2 /*return*/, chapterText];
                 }
             });
         });
     };
-    KolNovel.prototype.searchNovels = function (searchTerm, pageNo) {
+    PandaMTL.prototype.searchNovels = function (searchTerm, pageNo) {
         return __awaiter(this, void 0, void 0, function () {
             var url, headers, result, body, loadedCheerio, novels;
             return __generator(this, function (_a) {
@@ -318,7 +289,7 @@ var KolNovel = /** @class */ (function () {
             });
         });
     };
-    KolNovel.prototype.fetchImage = function (url) {
+    PandaMTL.prototype.fetchImage = function (url) {
         return __awaiter(this, void 0, void 0, function () {
             return __generator(this, function (_a) {
                 switch (_a.label) {
@@ -328,6 +299,6 @@ var KolNovel = /** @class */ (function () {
             });
         });
     };
-    return KolNovel;
+    return PandaMTL;
 }());
-exports.default = new KolNovel();
+exports.default = new PandaMTL();
