+++ plugins/english/pandaMTL.ts
@@ -3,11 +3,11 @@ import { fetchApi, fetchFile } from "@libs/fetch";
 import { Filter, FilterInputs } from "@libs/filterInputs";
 import { Plugin } from "@typings/plugin";
 
-class KolNovel implements Plugin.PluginBase {
-    id = "kolnovel";
-    name = "KolNovel";
-    icon = "multisrc/wpmangastream/icons/kolnovel.png"
-    site = "https://kolnovel.lol/";
+class PandaMTL implements Plugin.PluginBase {
+    id = "pandamtl";
+    name = "PandaMTL";
+    icon = "src/en/wordpress/icon.png";
+    site = "https://pandamtl.com/";
     version = "1.0.0";
     userAgent = "";
     cookieString = "";
@@ -25,6 +25,7 @@ class KolNovel implements Plugin.PluginBase {
         link += "&status=" + (filters?.status ? filters.status : "");
 
         link += "&order=" + (filters?.order ? filters.order : "popular");
+
         const headers = new Headers();
         if(this.cookieString){
             headers.append("cookie", this.cookieString);
@@ -62,7 +63,7 @@ class KolNovel implements Plugin.PluginBase {
         if(this.cookieString){
             headers.append("cookie", this.cookieString);
         }
-        const result = await fetch(url, {headers});
+        const result = await fetchApi(url, {headers});
         const body = await result.text();
 
         let loadedCheerio = parseHTML(body);
@@ -78,19 +79,18 @@ class KolNovel implements Plugin.PluginBase {
             loadedCheerio("img.wp-post-image").attr("data-src") ||
             loadedCheerio("img.wp-post-image").attr("src");
 
-        loadedCheerio("div.serl:nth-child(3) > span").each(function () {
+        loadedCheerio(".serl:nth-child(3) > span").each(function () {
             const detailName = loadedCheerio(this).text().trim();
             const detail = loadedCheerio(this).next().text().trim();
 
             switch (detailName) {
-                case "الكاتب":
                 case "Author":
                     novel.author = detail;
                     break;
             }
         });
 
-        novel.status = loadedCheerio("div.sertostat > span").attr("class");
+        novel.status = loadedCheerio(".sertostat > span").attr("class");
 
         novel.genres = loadedCheerio(".sertogenre")
             .children("a")
@@ -139,18 +139,13 @@ class KolNovel implements Plugin.PluginBase {
         if(this.cookieString){
             headers.append("cookie", this.cookieString);
         }
-        const result = await fetch(chapterUrl, {headers});
+        const result = await fetchApi(chapterUrl, {headers});
         const body = await result.text();
-    
+
         const loadedCheerio = parseHTML(body);
 
-        loadedCheerio('.epcontent > div, i').remove();
-        let ignore = loadedCheerio('article > style').text().trim().split(',');
-        ignore.push(...(ignore.pop()?.match(/^\.\w+/) || []));
-        ignore.map(tag => loadedCheerio(`p${tag}`).remove());
+        let chapterText = loadedCheerio(".epcontent").html() || "";
 
-        let chapterText = loadedCheerio('.epcontent').html() || "";
-    
         return chapterText;
     }
     async searchNovels(searchTerm: string, pageNo?: number | undefined): Promise<Plugin.NovelItem[]> {
@@ -186,140 +181,96 @@ class KolNovel implements Plugin.PluginBase {
     async fetchImage(url: string): Promise<string | undefined> {
         return await fetchFile(url);
     }
+
+
     filters = [
         {
             key: "order",
-            label: "ترتيب حسب",
+            label: "Sort By",
             values: [
-                { label: "الإعداد الأولي", value: "" },
-    
+                { label: "Default", value: "" },
+
                 { label: "A-Z", value: "title" },
-    
+
                 { label: "Z-A", value: "titlereverse" },
-    
-                { label: "أخر التحديثات", value: "update" },
-    
-                { label: "أخر ما تم إضافته", value: "latest" },
-    
-                { label: "الرائجة", value: "popular" },
+
+                { label: "Latest Update", value: "update" },
+
+                { label: "Latest Added", value: "latest" },
+
+                { label: "Popular", value: "popular" },
             ],
             inputType: FilterInputs.Picker,
         },
         {
             key: "status",
-            label: "الحالة",
+            label: "Status",
             values: [
                 { label: "All", value: "" },
-    
+
                 { label: "Ongoing", value: "ongoing" },
-    
+
                 { label: "Hiatus", value: "hiatus" },
-    
+
                 { label: "Completed", value: "completed" },
             ],
             inputType: FilterInputs.Picker,
         },
         {
             key: "type",
-            label: "النوع",
+            label: "Type",
             values: [
-                { label: "إنجليزية", value: "english" },
-    
-                { label: "روايةلايت", value: "light-novel" },
-    
-                { label: "روايةويب", value: "web-novel" },
-    
-                { label: "صينية", value: "chinese" },
-    
-                { label: "عربية", value: "arabic" },
-    
-                { label: "كورية", value: "korean" },
-    
-                { label: "يابانية", value: "japanese" },
+                { label: "Light Novel (KR)", value: "light-novel-kr" },
+
+                { label: "Web Novel", value: "web-novel" },
             ],
             inputType: FilterInputs.Checkbox,
         },
         {
             key: "genres",
-            label: "تصنيف",
+            label: "Genres",
             values: [
-                { label: "Wuxia", value: "wuxia" },
-    
-                { label: "Xianxia", value: "xianxia" },
-    
-                { label: "XUANHUAN", value: "xuanhuan" },
-    
-                { label: "أكشن", value: "action" },
-    
-                { label: "إثارة", value: "excitement" },
-    
-                { label: "إنتقالالىعالمأخر", value: "isekai" },
-    
-                { label: "إيتشي", value: "etchi" },
-    
-                { label: "الخيالالعلمي", value: "sci-fi" },
-    
-                { label: "بوليسي", value: "policy" },
-    
-                { label: "تاريخي", value: "historical" },
-    
-                { label: "تحقيقات", value: "%d8%aa%d8%ad%d9%82%d9%8a%d9%82" },
-    
-                { label: "تقمصشخصيات", value: "rpg" },
-    
-                { label: "جريمة", value: "crime" },
-    
-                { label: "جوسى", value: "josei" },
-    
-                { label: "حريم", value: "harem" },
-    
-                { label: "حياةمدرسية", value: "school-life" },
-    
-                { label: "خيالي(فانتازيا)", value: "fantasy" },
-    
-                { label: "دراما", value: "drama" },
-    
-                { label: "رعب", value: "horror" },
-    
-                { label: "رومانسي", value: "romantic" },
-    
-                { label: "سحر", value: "magic" },
-    
-                { label: "سينن", value: "senen" },
-    
-                { label: "شريحةمنالحياة", value: "slice-of-life" },
-    
-                { label: "شوجو", value: "shojo" },
-    
-                { label: "شونين", value: "shonen" },
-    
-                { label: "طبي", value: "medical" },
-    
-                { label: "ظواهرخارقةللطبيعة", value: "supernatural" },
-    
-                { label: "غموض", value: "mysteries" },
-    
-                { label: "فنونالقتال", value: "martial-arts" },
-    
-                { label: "قوىخارقة", value: "superpower" },
-    
-                { label: "كوميدي", value: "comedy" },
-    
-                { label: "مأساوي", value: "tragedy" },
-    
-                { label: "مابعدالكارثة", value: "after-the-disaster" },
-    
-                { label: "مغامرة", value: "adventure" },
-    
-                { label: "ميكا", value: "mechanical" },
-    
-                { label: "ناضج", value: "mature" },
-    
-                { label: "نفسي", value: "psychological" },
+                { label: "Action", value: "action" },
+
+                { label: "Adult", value: "adult" },
+
+                { label: "Adventure", value: "adventure" },
+
+                { label: "Comedy", value: "comedy" },
+
+                { label: "Ecchi", value: "ecchi" },
+
+                { label: "Fantasy", value: "fantasy" },
+
+                { label: "Harem", value: "harem" },
+
+                { label: "Josei", value: "josei" },
+
+                { label: "Martial Arts", value: "martial-arts" },
+
+                { label: "Mature", value: "mature" },
+
+                { label: "Romance", value: "romance" },
+
+                { label: "School Life", value: "school-life" },
+
+                { label: "Sci-fi", value: "sci-fi" },
+
+                { label: "Seinen", value: "seinen" },
+
+                { label: "Slice of Life", value: "slice-of-life" },
+
+                { label: "Smut", value: "smut" },
+
+                { label: "Sports", value: "sports" },
+
+                { label: "Supernatural", value: "supernatural" },
+
+                { label: "Tragedy", value: "tragedy" },
             ],
             inputType: FilterInputs.Checkbox,
         },
     ];
 }
 
-export default new KolNovel();
+export default new PandaMTL();
